perf: smaller browser bundle
This removes lodash for a significantly smaller bundle in the browser. From https://bundlephobia.com/result?p=ip-address@6.2.0 you can see that lodash is a significant portion of the bundle size.

The functions that were delegated to lodash are supported in native javascript by the majority of runtimes.

BREAKING CHANGE: Removes support for IE11, Android webview and Node.js < 8.0.0